LayoutTests/imported/w3c/web-platform-tests/css/css-align/blocks/align-content-block-006-expected.txt

FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
Show Text


PASS .test 1: start
FAIL .test 2: center assert_equals:
<div class="test" style="align-content: center" title="center">
    <div class="float" data-offset-y="15">FLT</div>
    <div class="in-flow" data-offset-y="25"></div>
    <div class="in-flow">
      <div class="float" data-offset-y="25">FLT</div>
      <span class="label">CENTER</span>
      <span class="abspos">ABS</span>
      <span class="relpos">REL</span>
      <div class="overflow">OVERFLOW</div>
    </div>
  </div>
offsetTop expected 15 but got 25
FAIL .test 3: end assert_equals:
<div class="test" style="align-content: end" title="end">
    <div class="float" data-offset-y="25">FLT</div>
    <div class="in-flow" data-offset-y="35"></div>
    <div class="in-flow">
      <div class="float" data-offset-y="35">FLT</div>
      <span class="label">END</span>
      <span class="abspos">ABS</span>
      <span class="relpos">REL</span>
      <div class="overflow">OVERFLOW</div>
    </div>
  </div>
offsetTop expected 25 but got 45
PASS .test 4: baseline
FAIL .test 5: last baseline assert_equals:
<div class="test" style="align-content: last baseline" title="last baseline">
    <div class="float" data-offset-y="25">FLT</div>
    <div class="in-flow" data-offset-y="35"></div>
    <div class="in-flow">
      <div class="float" data-offset-y="35">FLT</div>
      <span class="label">LAST BASELINE</span>
      <span class="abspos">ABS</span>
      <span class="relpos">REL</span>
      <div class="overflow">OVERFLOW</div>
    </div>
  </div>
offsetTop expected 25 but got 45
PASS .test 6: flex-start
FAIL .test 7: flex-end assert_equals:
<div class="test" style="align-content: flex-end" title="flex-end">
    <div class="float" data-offset-y="25">FLT</div>
    <div class="in-flow" data-offset-y="35"></div>
    <div class="in-flow">
      <div class="float" data-offset-y="35">FLT</div>
      <span class="label">FLEX-END</span>
      <span class="abspos">ABS</span>
      <span class="relpos">REL</span>
      <div class="overflow">OVERFLOW</div>
    </div>
  </div>
offsetTop expected 25 but got 45
PASS .test 8: unsafe start
FAIL .test 9: unsafe center assert_equals:
<div class="test" style="align-content: unsafe center" title="unsafe center">
    <div class="float" data-offset-y="15">FLT</div>
    <div class="in-flow" data-offset-y="25"></div>
    <div class="in-flow">
      <div class="float" data-offset-y="25">FLT</div>
      <span class="label">UNSAFE CENTER</span>
      <span class="abspos">ABS</span>
      <span class="relpos">REL</span>
      <div class="overflow">OVERFLOW</div>
    </div>
  </div>
offsetTop expected 15 but got 25
FAIL .test 10: unsafe end assert_equals:
<div class="test" style="align-content: unsafe end" title="unsafe end">
    <div class="float" data-offset-y="25">FLT</div>
    <div class="in-flow" data-offset-y="35"></div>
    <div class="in-flow">
      <div class="float" data-offset-y="35">FLT</div>
      <span class="label">UNSAFE END</span>
      <span class="abspos">ABS</span>
      <span class="relpos">REL</span>
      <div class="overflow">OVERFLOW</div>
    </div>
  </div>
offsetTop expected 25 but got 45
PASS .test 11: safe start
FAIL .test 12: safe center assert_equals:
<div class="test" style="align-content: safe center" title="safe center">
    <div class="float" "="" data-offset-y="15">FLT</div>
    <div class="in-flow" "="" data-offset-y="25"></div>
    <div class="in-flow">
      <div class="float" "="" data-offset-y="25">FLT</div>
      <span class="label">SAFE CENTER</span>
      <span class="abspos">ABS</span>
      <span class="relpos">REL</span>
      <div class="overflow">OVERFLOW</div>
    </div>
  </div>
offsetTop expected 15 but got 25
FAIL .test 13: safe end assert_equals:
<div class="test" style="align-content: safe end" title="safe end">
    <div class="float" data-offset-y="25">FLT</div>
    <div class="in-flow" data-offset-y="35"></div>
    <div class="in-flow">
      <div class="float" data-offset-y="35">FLT</div>
      <span class="label">SAFE END</span>
      <span class="abspos">ABS</span>
      <span class="relpos">REL</span>
      <div class="overflow">OVERFLOW</div>
    </div>
  </div>
offsetTop expected 25 but got 45
FAIL .test 14: space-evenly assert_equals:
<div class="test" style="align-content: space-evenly" title="space-evenly">
    <div class="float" data-offset-y="15">FLT</div>
    <div class="in-flow" data-offset-y="25"></div>
    <div class="in-flow">
      <div class="float" data-offset-y="25">FLT</div>
      <span class="label">SPACE-EVENLY</span>
      <span class="abspos">ABS</span>
      <span class="relpos">REL</span>
      <div class="overflow">OVERFLOW</div>
    </div>
  </div>
offsetTop expected 15 but got 25
PASS .test 15: space-between
FAIL .test 16: space-around assert_equals:
<div class="test" style="align-content: space-around" title="space-around">
    <div class="float" data-offset-y="15">FLT</div>
    <div class="in-flow" data-offset-y="25"></div>
    <div class="in-flow">
      <div class="float" data-offset-y="25">FLT</div>
      <span class="label">SPACE-AROUND</span>
      <span class="abspos">ABS</span>
      <span class="relpos">REL</span>
      <div class="overflow">OVERFLOW</div>
    </div>
  </div>
offsetTop expected 15 but got 25
PASS .test 17: normal
